hej hej hej

+++ RERPAPI.IRepo/IGenericRepo.cs
@@ -9,12 +9,12 @@ namespace RERPAPI.IRepo
     public interface IGenericRepo<T> where T : class
     {
         List<T> GetAll();
-        T GetByID(long id);
+        T GetByID(int id);
 
         int Create(T item);
         int CreateRange(List<T> items);
         int Update(T item);
-        int Delete(long id);
+        int Delete(int id);
         int DeleteRange(List<T> items);
 
         Task<int> CreateAsync(T item);
